Switch to using stdint.h

+++ src/hash.c
@@ -1,12 +1,12 @@
 #include "hash.h"
 
-unsigned long hash(const char* str, unsigned long len) {
-    unsigned long hash = 5381;
-    for (unsigned long i = 0; i < len; i++)
+uint64_t hash(const char* str, uint64_t len) {
+    uint32_t hash = 5381;
+    for (uint64_t i = 0; i < len; i++)
         hash = hash * 33 + str[i];
     return hash;
 }
 
-unsigned long combine_hash(unsigned long hash_a, unsigned long hash_b) {
+uint64_t combine_hash(uint64_t hash_a, uint64_t hash_b) {
     return (5381 + hash_a) << 5 + hash_b;
 }
